backend/app/services/celery_tasks.py
"""
ABDSシステム - Celeryタスク（コメント実装）
非同期画像検索処理用のCeleryタスク

実際の本番環境では、以下のようにCeleryを統合します：

1. Celeryの設定:
   ```python
   # app/core/celery_app.py
   from celery import Celery
   
   celery_app = Celery(
       "abds_system",
       broker="redis://localhost:6379/0",
       backend="redis://localhost:6379/0",
       include=["app.services.celery_tasks"]
   )
   
   celery_app.conf.update(
       task_serializer="json",
       accept_content=["json"],
       result_serializer="json",
       timezone="UTC",
       enable_utc=True,
       task_track_started=True,
       task_time_limit=30 * 60,  # 30分
       task_soft_time_limit=25 * 60,  # 25分
       worker_prefetch_multiplier=1,
       worker_max_tasks_per_child=1000,
   )
   ```

2. タスクの実装:
"""

# from celery import Celery, current_task
# from celery.exceptions import Retry
# import logging
# from datetime import datetime
# from typing import Dict, Any
# 
# from app.core.celery_app import celery_app
# from app.core.database import SessionLocal
# from app.models import Image, SearchResult as DBSearchResult, ImageStatus, ThreatLevel
# from app.services.image_search import create_image_search_service, SearchAPIError, RateLimitExceededError
# 
# logger = logging.getLogger(__name__)


# @celery_app.task(bind=True, name="search_similar_images")
# def search_similar_images_task(
#     self,
#     image_id: str,
#     service_type: str = "serpapi",
#     max_results: int = 10
# ) -> Dict[str, Any]:
#     """
#     非同期画像検索タスク
#     
#     Args:
#         self: Celeryタスクインスタンス
#         image_id: 画像ID
#         service_type: 検索サービスタイプ ("google" または "serpapi")
#         max_results: 最大結果数
#         
#     Returns:
#         タスク結果辞書
#         
#     Raises:
#         Retry: 再試行可能なエラーの場合
#     """
#     db = None
#     
#     try:
#         # データベースセッション作成
#         db = SessionLocal()
#         
#         # 画像情報を取得
#         image = db.query(Image).filter(Image.id == image_id).first()
#         if not image:
#             raise ValueError(f"画像が見つかりません: {image_id}")
#         
#         # タスクステータス更新
#         self.update_state(
#             state="PROGRESS",
#             meta={"progress": 0.1, "message": "検索サービス初期化中..."}
#         )
#         
#         # 検索サービス作成
#         search_service = create_image_search_service(service_type)
#         
#         # タスクステータス更新
#         self.update_state(
#             state="PROGRESS", 
#             meta={"progress": 0.3, "message": "画像検索実行中..."}
#         )
#         
#         # 画像検索実行
#         logger.info(f"画像検索開始: image_id={image_id}, service={service_type}")
#         
#         # 非同期関数を同期的に実行
#         import asyncio
#         try:
#             loop = asyncio.get_event_loop()
#         except RuntimeError:
#             loop = asyncio.new_event_loop()
#             asyncio.set_event_loop(loop)
#         
#         search_results = loop.run_until_complete(
#             search_service.search_similar_images(
#                 image_path=image.file_path,
#                 max_results=max_results
#             )
#         )
#         
#         # タスクステータス更新
#         self.update_state(
#             state="PROGRESS",
#             meta={"progress": 0.7, "message": "結果をデータベースに保存中..."}
#         )
#         
#         # 結果をデータベースに保存
#         saved_count = 0
#         for result in search_results:
#             try:
#                 # 重複チェック
#                 existing = db.query(DBSearchResult).filter(
#                     DBSearchResult.image_id == image_id,
#                     DBSearchResult.found_url == result.url
#                 ).first()
#                 
#                 if not existing:
#                     db_result = DBSearchResult(
#                         image_id=image_id,
#                         found_url=result.url,
#                         domain=result.source_domain,
#                         similarity_score=result.similarity_score,
#                         is_official=False,  # TODO: 公式サイト判定ロジック実装
#                         threat_level=ThreatLevel.SAFE,  # TODO: 脅威レベル判定実装
#                         analyzed_at=datetime.utcnow()
#                     )
#                     db.add(db_result)
#                     saved_count += 1
#                 
#             except Exception as e:
#                 logger.warning(f"検索結果の保存でエラー: {e}")
#                 continue
#         
#         # データベースコミット
#         db.commit()
#         
#         # 画像ステータス更新
#         image.status = ImageStatus.COMPLETED
#         db.commit()
#         
#         # タスク完了
#         result_data = {
#             "status": "completed",
#             "image_id": image_id,
#             "service_type": service_type,
#             "total_results": len(search_results),
#             "saved_results": saved_count,
#             "completed_at": datetime.utcnow().isoformat(),
#             "progress": 1.0
#         }
#         
#         logger.info(f"画像検索完了: {result_data}")
#         return result_data
#         
#     except RateLimitExceededError as e:
#         logger.warning(f"レート制限エラー: {e}")
#         
#         # 画像ステータス更新
#         if image:
#             image.status = ImageStatus.FAILED
#             db.commit()
#         
#         # 60分後に再試行
#         raise self.retry(exc=e, countdown=3600, max_retries=3)
#         
#     except SearchAPIError as e:
#         logger.error(f"検索APIエラー: {e}")
#         
#         # 画像ステータス更新
#         if image:
#             image.status = ImageStatus.FAILED
#             db.commit()
#         
#         # 10分後に再試行
#         raise self.retry(exc=e, countdown=600, max_retries=2)
#         
#     except Exception as e:
#         logger.error(f"画像検索タスクでエラー: {e}")
#         
#         # 画像ステータス更新
#         if image:
#             image.status = ImageStatus.FAILED
#             db.commit()
#         
#         # 一般的なエラーは3分後に再試行
#         raise self.retry(exc=e, countdown=180, max_retries=1)
#         
#     finally:
#         if db:
#             db.close()


# @celery_app.task(name="bulk_search_images")
# def bulk_search_images_task(
#     image_ids: list,
#     service_type: str = "serpapi",
#     max_results_per_image: int = 10
# ) -> Dict[str, Any]:
#     """
#     一括画像検索タスク
#     
#     Args:
#         image_ids: 画像IDのリスト
#         service_type: 検索サービスタイプ
#         max_results_per_image: 画像あたりの最大結果数
#         
#     Returns:
#         一括処理結果
#     """
#     results = []
#     failed_images = []
#     
#     for image_id in image_ids:
#         try:
#             # 個別のタスクを実行
#             task_result = search_similar_images_task.apply_async(
#                 args=[image_id, service_type, max_results_per_image]
#             )
#             
#             # 結果を待機（タイムアウト: 10分）
#             result = task_result.get(timeout=600)
#             results.append(result)
#             
#         except Exception as e:
#             logger.error(f"画像 {image_id} の検索でエラー: {e}")
#             failed_images.append({"image_id": image_id, "error": str(e)})
#     
#     return {
#         "status": "completed" if not failed_images else "partial",
#         "total_images": len(image_ids),
#         "successful_searches": len(results),
#         "failed_searches": len(failed_images),
#         "results": results,
#         "failures": failed_images,
#         "completed_at": datetime.utcnow().isoformat()
#     }


# @celery_app.task(name="cleanup_old_search_results")
# def cleanup_old_search_results_task(days_old: int = 30) -> Dict[str, Any]:
#     """
#     古い検索結果のクリーンアップタスク
#     
#     Args:
#         days_old: 削除対象の日数（デフォルト: 30日）
#         
#     Returns:
#         クリーンアップ結果
#     """
#     db = None
#     
#     try:
#         db = SessionLocal()
#         
#         # 指定日数より古い検索結果を削除
#         cutoff_date = datetime.utcnow() - timedelta(days=days_old)
#         
#         deleted_count = db.query(DBSearchResult).filter(
#             DBSearchResult.analyzed_at < cutoff_date
#         ).delete()
#         
#         db.commit()
#         
#         logger.info(f"古い検索結果のクリーンアップ完了: {deleted_count}件削除")
#         
#         return {
#             "status": "completed",
#             "deleted_count": deleted_count,
#             "cutoff_date": cutoff_date.isoformat(),
#             "completed_at": datetime.utcnow().isoformat()
#         }
#         
#     except Exception as e:
#         logger.error(f"クリーンアップタスクでエラー: {e}")
#         return {
#             "status": "failed",
#             "error": str(e),
#             "completed_at": datetime.utcnow().isoformat()
#         }
#         
#     finally:
#         if db:
#             db.close()


"""
Celeryワーカーの起動方法:

1. Redisサーバーの起動:
   ```bash
   redis-server
   ```

2. Celeryワーカーの起動:
   ```bash
   celery -A app.core.celery_app worker --loglevel=info
   ```

3. Celery Beatの起動（定期タスク用）:
   ```bash
   celery -A app.core.celery_app beat --loglevel=info
   ```

4. Celery Flowerの起動（監視UI）:
   ```bash
   celery -A app.core.celery_app flower
   ```

定期タスクの設定例:

```python
# app/core/celery_app.py に追加
from celery.schedules import crontab

celery_app.conf.beat_schedule = {
    'cleanup-old-search-results': {
        'task': 'cleanup_old_search_results',
        'schedule': crontab(hour=2, minute=0),  # 毎日午前2時
        'args': (30,)  # 30日以上古い結果を削除
    },
}
```

API統合例:

```python
# FastAPIエンドポイントでの使用
@router.post("/search/async/{image_id}")
async def start_async_search(image_id: UUID):
    task = search_similar_images_task.delay(str(image_id))
    return {"task_id": task.id, "status": "started"}

@router.get("/search/task/{task_id}")
async def get_task_status(task_id: str):
    task = search_similar_images_task.AsyncResult(task_id)
    return {
        "task_id": task_id,
        "status": task.status,
        "result": task.result if task.ready() else None
    }
```
"""
